==> notebooks/adhd_beliefs_pt/01-data_preparation.ipynb <==
# Formatting Dataset

import pandas as pd
from datetime import datetime
import pytz

file_path = "../../data/adhd-beliefs-pt/adhd-beliefs-pt.csv"
df = pd.read_csv(file_path)
df.head()

df.columns = [
    "datetime",
    "consent",
    "sex",
    "adhd_diagnosis",
    "age",
    "education",
    "occupation",
    "dialect",
    "forgetting_objects",
    "forgetting_responsabilities",
    "emotion_management",
    "emotion_reactions",
    "emotion_choices",
    "emotion_intense_reaction",
    "strong_mood_swings",
    "control_mood_swings",
    "mood_swings_outside_consequence",
    "justice_sense",
    "defend_beliefs",
    "express_writing",
    "topic_change_involuntary",
    "related_topic",
    "topic_change_unrealized",
    "parallel_topic",
    "nonlinear_storytelling",
    "excessive_details",
    "forgetting_mid_conversation",
    "talkback",
    "talkback_authority",
    "talkback_comfortable",
    "interrupting_involuntary",
    "excited_opinion",
    "want_interrupt",
    "want_interrupt_control",
    "called_out_expressing",
    "called_out_exalted",
    "called_out_talking_loud",
    "called_out_nonlinear_talk",
    "called_out_nonlinear_write",
    "called_out_gestures",
    "called_out_talking",
    "detailed_opinions",
    "detailed_opinions_interest",
    "detailed_opinions_people",
    "detailed_opinions_misunderstood",
    "swearing",
    "swearing_anywhere",
    "swearing_casual",
    "talk_fast",
    "talk_fast_unrealized",
    "talk_fast_uncontrolled",
    "called_out_talk_fast",
    "need_fast_talk_interest",
    "need_fast_talk_information",
    "speaking_before_thinking",
    "something_to_add",
    "something_to_add_timid",
    "something_to_add_impulsive",
    "special_interest",
    "diary_entry",
    "selfdefining_memory",
    "empty_sheet",
]

df

df = df.drop_duplicates()
df.drop(columns=["consent"], inplace=True)

def parse_portuguese_datetime(s: str) -> str:
    s_clean = s.replace(" da manhã", " AM").replace(" da tarde", " PM")
    dt_naive = datetime.strptime(s_clean, "%Y/%m/%d %I:%M:%S %p CET")
    cet = pytz.timezone('CET')
    dt_cet = cet.localize(dt_naive)
    return dt_cet.isoformat() # ISO 8601

df['datetime'] = df['datetime'].apply(parse_portuguese_datetime)

for col in df.columns:
    if col not in [
        "age",
        "special_interest",
        "diary_entry",
        "selfdefining_memory",
        "empty_sheet",
    ]:
        df[col] = df[col].astype("category")

encoding = {"Nunca": 1, "Raramente": 2, "Às vezes": 3, "Frequentemente": 4, "Sempre": 5}

for col in df.columns:
    if col not in [
        "datetime",
        "sex",
        "adhd_diagnosis",
        "age",
        "education",
        "occupation",
        "dialect",
        "special_interest",
        "diary_entry",
        "selfdefining_memory",
        "empty_sheet"
    ]:
        df[col] = df[col].map(encoding)

df

### Join Education Categories

df["education"].unique()

# make entries in 'education' that are category 'bacharelado' into 'licenciatura' and 'mestre' into 'mestrado'
df["education"] = df["education"].astype("category")
df["education"] = df["education"].replace(
    {
        "Bacharelado": "Licenciatura",
        "Pôs ": "Pós-Graduação",
        "Pos-Graduação ": "Pós-Graduação",
    }
)

df["education"].unique()

### Join Occupation Categories

df["occupation"].unique()

df["occupation"] = df["occupation"].astype("category")
df["occupation"] = df["occupation"].replace(
    {
        "Doméstica ": "Trabalhador",
        "Pós-graduação ": "Estudante",
        "Freelance": "Trabalhador",
        "estudante/bombeira voluntária": "Trabalhador-estudante",
    }
)

df["occupation"].unique()

### Saving to Local Disk

df.to_pickle("../../data/adhd-beliefs-pt/adhd-beliefs-pt-prepared.pkl")
