==> notebooks/adhd_beliefs_pt/04-data_preprocessing.ipynb <==
import nltk
import pandas as pd
import stanza
from nltk.corpus import stopwords

file_path = "../../data/adhd-beliefs-pt/adhd-beliefs-pt-anonymized.pkl"
df = pd.read_pickle(file_path)
df.head()

nltk.download("stopwords")
stopwords_pt = set(stopwords.words("portuguese"))

stanza.download("pt")
nlp = stanza.Pipeline(lang="pt", processors="tokenize,mwt,pos,lemma")

def preprocess_text(text: str) -> str:
    doc = nlp(text.lower())
    tokens = []
    for sentence in doc.sentences:
        for word in sentence.words:
            lemma = word.lemma
            if lemma.isalpha() and lemma not in stopwords_pt:
                tokens.append(lemma)
    return " ".join(tokens)

text_columns = ["special_interest", "diary_entry", "selfdefining_memory", "empty_sheet"]

for column in text_columns:
    df[column] = df[column].astype(str)
    df[column] = df[column].apply(preprocess_text)

temp_df = df[text_columns].replace("nan", "").fillna("")
df["merged_text"] = temp_df.agg(" ".join, axis=1)
df["merged_text"] = df["merged_text"].replace("   ", "nan")

file_path = "../../data/adhd-beliefs-pt/adhd-beliefs-pt-preprocessed.pkl"
df.to_pickle(file_path)
print("Text preprocessing completed and saved to", file_path)

df
