==> notebooks/adhd_beliefs_pt/10-prepare_clustering.ipynb <==
import numpy as np
import pandas as pd

df = pd.read_pickle("../../data/adhd-beliefs-pt/adhd-beliefs-pt-embeddings-serafim.pkl")
df

question_types = ["special_interest", "diary_entry", "selfdefining_memory", "empty_sheet"]
topic_df = pd.concat([df.assign(question=q) for q in question_types], ignore_index=False)
topic_df["response"] = topic_df.apply(lambda row: row[row["question"]], axis=1)
topic_df["response_embedding"] = topic_df.apply(lambda row: row[f"{row['question']}_embedding"], axis=1)
topic_df = topic_df.drop(columns=["special_interest", "diary_entry", "selfdefining_memory", "empty_sheet", "merged_text", "special_interest_embedding", "diary_entry_embedding", "selfdefining_memory_embedding", "empty_sheet_embedding", "merged_text_embedding"])
topic_df = topic_df.dropna(subset=["response"])
topic_df

question_counts = topic_df["question"].value_counts()
question_counts

topic_df['group'] = (
    topic_df['sex'].map({'Feminino':'Female','Masculino':'Male'}).astype(str) + '_' +
    np.where(topic_df['adhd_diagnosis']=="Sim, diagnosticado", 'ADHD', 'noADHD')
)
groups = topic_df['group'].unique().tolist()
group_counts = topic_df['group'].value_counts()
group_counts

topic_df

file_path = "../../data/adhd-beliefs-pt/adhd-beliefs-pt-embeddings-serafim-bertopic.pkl"
topic_df.to_pickle(file_path)
